+++ crates/consensus/beacon/src/engine/error.rs
@@ -1,16 +1,19 @@
 use reth_stages::PipelineError;
 use thiserror::Error;
 
+/// Beacon engine result.
+pub type BeaconEngineResult<Ok> = Result<Ok, BeaconEngineError>;
+
 /// The error wrapper for the beacon consensus engine.
 #[derive(Error, Debug)]
 pub enum BeaconEngineError {
     /// Pipeline error.
-    #[error("Controller encountered a pipeline error.")]
+    #[error(transparent)]
     Pipeline(#[from] PipelineError),
     /// Database error.
-    #[error("Controller encountered a database error.")]
+    #[error(transparent)]
     Database(#[from] reth_db::Error),
     /// Common error. Wrapper around [reth_interfaces::Error].
-    #[error("Controller encountered an error.")]
+    #[error(transparent)]
     Common(#[from] reth_interfaces::Error),
 }
