+++ crates/rpc/rpc-engine-api/Cargo.toml
@@ -11,10 +11,10 @@ description = "Implementation of Engine API"
 reth-primitives = { path = "../../primitives" }
 reth-interfaces = { path = "../../interfaces" }
 reth-provider = { path = "../../storage/provider" }
-reth-rlp = { path = "../../rlp" }
 reth-executor = { path = "../../executor" }
 reth-revm = { path = "../../revm" }
 reth-rpc-types = { path = "../rpc-types" }
+reth-beacon-consensus = { path = "../../consensus/beacon" }
 
 # async
 futures = "0.3"
