+++ crates/rpc/rpc-types/Cargo.toml
@@ -27,5 +27,9 @@ jsonrpsee-types = { version = "0.16" }
 lru = "0.9"
 
 [dev-dependencies]
-rand = "0.8"
+# reth
 reth-interfaces = { path = "../../interfaces", features = ["test-utils"] }
+
+# misc
+rand = "0.8"
+assert_matches = "1.5"
